[BACKLOG-12101] Extracts TestQuartzScheduler from JaxWsSchedulerServiceIT for use in other modules
