+++ scheduler/src/test/java/org/pentaho/platform/scheduler2/ws/test/TestQuartzScheduler.java
@@ -1,0 +1,32 @@
+/*!
+ * This program is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License, version 2.1 as published by the Free Software
+ * Foundation.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along with this
+ * program; if not, you can obtain a copy at http://www.gnu.org/licenses/old-licenses/lgpl-2.1.html
+ * or from the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * Copyright (c) 2002-2016 Pentaho Corporation..  All rights reserved.
+ */
+
+package org.pentaho.platform.scheduler2.ws.test;
+
+import org.pentaho.platform.engine.security.SecurityHelper;
+import org.pentaho.platform.scheduler2.quartz.QuartzScheduler;
+
+public class TestQuartzScheduler extends QuartzScheduler {
+
+  static final String TEST_USER = "TestUser";
+
+  @Override
+  protected String getCurrentUser() {
+    SecurityHelper.getInstance().becomeUser( TEST_USER );
+    return super.getCurrentUser();
+  }
+}
